server/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.2
zipp==3.15.0
